# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ApiError:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  dto.ApiResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ApiError'
      meta:
        $ref: '#/definitions/dto.ResponseMeta'
      success:
        type: boolean
    type: object
...
  dto.AvailabilityWindow:
    properties:
      nextWeek:
        type: number
      thisWeek:
        type: number
    type: object
...
  dto.IndividualWorkloadResponse:
    properties:
      currentAllocation:
        $ref: '#/definitions/dto.MemberAllocation'
      historicalTrend:
        items:
          $ref: '#/definitions/dto.UtilizationTrend'
        type: array
      name:
        type: string
      personId:
        type: string
      tasks:
        items:
          $ref: '#/definitions/dto.TaskAllocation'
        type: array
      upcomingWeeks:
        items:
          $ref: '#/definitions/dto.WeeklyForecast'
        type: array
    type: object
  dto.MemberAllocation:
    properties:
      assignedHours:
        type: number
      capacityHours:
        type: number
      percentage:
        type: integer
    type: object
...
  dto.ProjectAllocation:
    properties:
      percentage:
        type: number
      projectId:
        type: string
      projectName:
        type: string
      totalHours:
        type: number
    type: object
...
  dto.RebalanceSuggestion:
    properties:
      currentOwner:
        type: string
      impact:
        type: string
      reason:
        type: string
      suggestedOwner:
        type: string
      taskId:
        type: string
      taskTitle:
        type: string
    type: object
  dto.RebalanceWorkloadRequest:
    properties:
      maxUtilization:
        maximum: 120
        minimum: 80
        type: integer
      personId:
        type: string
      projectId:
        type: string
      targetDate:
        type: string
    type: object
  dto.RebalanceWorkloadResponse:
    properties:
      suggestions:
        items:
          $ref: '#/definitions/dto.RebalanceSuggestion'
        type: array
      totalImpact:
        type: string
    type: object
...
  dto.ResponseMeta:
    properties:
      hasMore:
        type: boolean
      nextCursor:
        type: string
      page:
        type: integer
      perPage:
        type: integer
      requestId:
        type: string
      timestamp:
        type: string
      total:
        type: integer
    type: object
  dto.RiskPeriod:
    properties:
      endWeek:
        type: string
      reason:
        type: string
      severity:
        type: string
      startWeek:
        type: string
    type: object
...
  dto.TaskAllocation:
    properties:
      allocationThisWeek:
        type: number
      estimatedHours:
        type: number
      projectId:
        type: string
      taskId:
        type: string
      title:
        type: string
    type: object
...
  dto.TeamMemberWorkload:
    properties:
      allocation:
        $ref: '#/definitions/dto.MemberAllocation'
      availability:
        $ref: '#/definitions/dto.AvailabilityWindow'
      name:
        type: string
      personId:
        type: string
      riskLevel:
        type: string
      role:
        type: string
      status:
        type: string
      tasks:
        items:
          $ref: '#/definitions/dto.TaskAllocation'
        type: array
    type: object
  dto.TeamWorkloadResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/dto.TeamMemberWorkload'
        type: array
      summary:
        $ref: '#/definitions/dto.WorkloadSummary'
      teamAllocation:
        type: number
      teamCapacity:
        type: number
      utilizationRate:
        type: number
      weekStarting:
        type: string
    type: object
...
  dto.UtilizationTrend:
    properties:
      date:
        type: string
      optimalCount:
        type: integer
      overCount:
        type: integer
      underCount:
        type: integer
      utilization:
        type: number
    type: object
...
  dto.WeeklyForecast:
    properties:
      allocation:
        type: integer
      assignedHours:
        type: number
      risk:
        type: string
      tasks:
        type: integer
      weekStarting:
        type: string
    type: object
  dto.WorkloadAnalyticsResponse:
    properties:
      avgUtilization:
        type: number
      byProject:
        items:
          $ref: '#/definitions/dto.ProjectAllocation'
        type: array
      distribution:
        $ref: '#/definitions/dto.WorkloadDistribution'
      peakUtilization:
        type: number
      period:
        type: string
      trends:
        items:
          $ref: '#/definitions/dto.UtilizationTrend'
        type: array
    type: object
  dto.WorkloadDistribution:
    properties:
      available:
        type: integer
      optimal:
        type: integer
      overallocated:
        type: integer
      underutilized:
        type: integer
    type: object
  dto.WorkloadForecastResponse:
    properties:
      forecast:
        items:
          $ref: '#/definitions/dto.WeeklyForecast'
        type: array
      personId:
        type: string
      recommendations:
        items:
          type: string
        type: array
      riskPeriods:
        items:
          $ref: '#/definitions/dto.RiskPeriod'
        type: array
    type: object
  dto.WorkloadSummary:
    properties:
      available:
        type: integer
      optimal:
        type: integer
      overallocated:
        type: integer
      underutilized:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Xephyr Project Management Intelligence API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Xephyr API
  version: "1.0"
paths:
...
  /workload/analytics:
    get:
      consumes:
      - application/json
      description: Get workload analytics for the organization
      parameters:
      - default: 30d
        description: Time period
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.WorkloadAnalyticsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get workload analytics
      tags:
      - workload
  /workload/forecast:
    get:
      consumes:
      - application/json
      description: Get workload forecast for a person
      parameters:
      - description: Person ID
        in: query
        name: personId
        required: true
        type: string
      - default: 8
        description: Number of weeks
        in: query
        name: weeks
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.WorkloadForecastResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get workload forecast
      tags:
      - workload
  /workload/people/{personId}:
    get:
      consumes:
      - application/json
      description: Get workload for a specific person
      parameters:
      - description: Person ID
        in: path
        name: personId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.IndividualWorkloadResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get individual workload
      tags:
      - workload
  /workload/rebalance:
    post:
      consumes:
      - application/json
      description: Get suggestions for rebalancing workload
      parameters:
      - description: Rebalance request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RebalanceWorkloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RebalanceWorkloadResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get rebalance suggestions
      tags:
      - workload
  /workload/team:
    get:
      consumes:
      - application/json
      description: Get team workload overview
      parameters:
      - description: Week starting date
        in: query
        name: week
        type: string
      - description: Include forecast
        in: query
        name: includeForecast
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TeamWorkloadResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get team workload
      tags:
      - workload
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"